+++ Examples/GUI/System_Network/Example_find_all_device/main.py
@@ -109,7 +109,6 @@ def main():
 
 if __name__ == "__main__":
     ## Create device handle 
-    dev = pywpc.Broadcaster()
+    dev = pywpc.DeviceFinder()
     main()
-